+++ flutter code/report/lib/take_picture_screen.dart
@@ -5,7 +5,6 @@ import 'package:path/path.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:report/bar.dart';
 import 'package:report/drawer.dart';
-import 'package:report/newBar.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 // ignore: must_be_immutable
